implementa votação do associado e suas regras para sessão de uma pauta

+++ contract/src/main/java/br/com/rjguastalli/v1/sessao/model/request/SessaoRequest.java
@@ -26,6 +26,6 @@ public class SessaoRequest {
 
     @ApiModelProperty(value = "Tempo que a sessão de votação ficará aberta.")
     @Builder.Default
-    private Integer tempoAbertura = 1;
+    private Long tempoAbertura = 1L;
 
 }
